Reorganized Test folder structure
Resolves #297

- Integration tests should now be located in the integration folder
- Differentaiton between unit and integration tests is done using JUnit categories
- Add maven profile all that executes unit and integration tests
- Remove outdated CategoryAwareRunners
- Add default IntelliJ Run Configurations
- Update Jenkinsfile to execute integration and unit tests in seperate stages
- Update travis.yml to run "mvn verify -P all" instead of "mvn test" (Travis will also run integration tests)
- Add ADR for test approach
- Add test guidelines
- Rename integration tests according to the test guidelines
- Add jacoco report generation for integration tests
