+++ docs/adr/index.md
@@ -16,6 +16,7 @@ This log lists the architectural decisions for *[project name]*.
 - [ADR-0009](0009-cloudfoundry-cli) - CloudFoundry CLI
 
 - [ADR-0010](0010-use-docker-java.md) - How to perform Kubernetes end to end tests
+- [ADR-0011](0011-put-integration-tests-in-seperate-folder.md) - Put Integration Tests in seperate source directory
 <!-- adrlogstop -->
 
 For new ADRs, please use [template.md](template.md) as basis.
