+++ docs/config/test-run-configurations/Readme.md
@@ -1,0 +1,13 @@
+# Test Run Configurations
+
+The Run Configurations in this Folder can be copied into th `.idea/runConfigurations/` folder to use. 
+
+They allow you to run:
+ - All Unit Tests of the server module
+ - All Integration Tests of the server module
+ - All Tests of the server module
+ - All Tests of the retrofit-wrapper module
+ - All Tests of the cli module
+ - All Tests (Coverage does not work!)
+ 
+These Configurations need IntelliJ Idea to run!
