+++ server/src/integration/java/org/opentosca/toscana/IntegrationTest.java
@@ -1,0 +1,4 @@
+package org.opentosca.toscana;
+
+public interface IntegrationTest {
+}
