+++ server/src/integration/java/org/opentosca/toscana/core/BaseJUnitIntegrationTest.java
@@ -1,0 +1,9 @@
+package org.opentosca.toscana.core;
+
+import org.opentosca.toscana.IntegrationTest;
+
+import org.junit.experimental.categories.Category;
+
+@Category(IntegrationTest.class)
+public abstract class BaseJUnitIntegrationTest extends BaseJUnitTest {
+}
