+++ server/src/test/java/org/opentosca/toscana/UnitTest.java
@@ -1,0 +1,4 @@
+package org.opentosca.toscana;
+
+public interface UnitTest {
+}
