+++ server/src/test/java/org/opentosca/toscana/core/BaseSpringTest.java
@@ -5,14 +5,11 @@ import java.io.IOException;
 
 import javax.annotation.PostConstruct;
 
-import org.opentosca.toscana.core.testutils.CategoryAwareSpringRunner;
 import org.opentosca.toscana.core.util.Preferences;
 
 import org.apache.commons.io.FileUtils;
 import org.junit.After;
-import org.junit.AfterClass;
 import org.junit.Before;
-import org.junit.BeforeClass;
 import org.junit.runner.RunWith;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -20,13 +17,14 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.TestPropertySource;
+import org.springframework.test.context.junit4.SpringRunner;
 
 /**
  Extend from this class in order to inherit important configurations Sets up Spring Test Context regarding to the Test
  Configuration After every test method, refreshes the context.
  If disk access is needed, use {@link #tmpdir}
  */
-@RunWith(CategoryAwareSpringRunner.class)
+@RunWith(SpringRunner.class)
 @ContextConfiguration(classes = {TestCoreConfiguration.class})
 @DirtiesContext(classMode = DirtiesContext.ClassMode.AFTER_EACH_TEST_METHOD)
 @TestPropertySource("classpath:test-properties.yml")
@@ -48,7 +46,6 @@ public abstract class BaseSpringTest extends BaseTest {
         tmpdir = preferences.getDataDir();
     }
 
-    
     @Before
     public final void startupPrepareDisk() throws IOException {
         FileUtils.deleteDirectory(tmpdir);
@@ -61,5 +58,4 @@ public abstract class BaseSpringTest extends BaseTest {
         FileUtils.deleteDirectory(tmpdir);
         logger.info("Cleaned up disk");
     }
-    
 }
