--- server/src/test/java/org/opentosca/toscana/core/testutils/TestCategory.java
@@ -1,15 +1,0 @@
-package org.opentosca.toscana.core.testutils;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Inherited;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-//Allow retention on class and Method
-@Target(value = {ElementType.METHOD, ElementType.TYPE})
-@Retention(RetentionPolicy.RUNTIME)
-@Inherited
-public @interface TestCategory {
-    TestCategories value() default TestCategories.FAST;
-}
